requirements.txt
# Core requirements for the API
fastapi>=0.95.1
uvicorn>=0.22.0
python-multipart>=0.0.6
ffmpeg-python>=0.2.0
pydantic>=1.10.7
requests>=2.28.2
python-dotenv>=1.0.0

# For speaker diarization
torch>=2.0.0
pyannote.audio>=3.1.0  # Required for speaker-diarization-3.1 model

# Linting Tools
black==23.12.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0
pylint==3.0.3
types-requests==2.31.0.20240311

# Kaggle
kagglehub>=0.1.0
